Refactor. Add base class for generators
