+++ lib/generators/mi/mi_generator.rb
@@ -6,20 +6,7 @@ require 'strscan'
 
 module Mi
   module Generators
-    class MiGenerator < Rails::Generators::Base
-      include Rails::Generators::Migration
-
-      def self.next_migration_number(dirname)
-        next_migration_number = current_migration_number(dirname) + 1
-        ActiveRecord::Migration.next_migration_number(next_migration_number)
-      end
-
-      Methods = {
-        '+' => 'add_column',
-        '-' => 'remove_column',
-        '%' => 'change_column',
-      }.freeze
-
+    class MiGenerator < Base
       source_root File.expand_path('../templates', __FILE__)
       namespace "mi"
 
@@ -35,42 +22,6 @@ module Mi
 
       private
 
-      def arguments
-        @_initializer[0..1].flatten
-      end
-
-      def arg_groups
-        @arg_groups ||= (
-          args = arguments.reject{|x| x.start_with?('--')}
-
-          current = nil
-          res = args.group_by do |a|
-            if %w[+ - %].include? a[0]
-              current
-            else
-              current = a
-              nil
-            end
-          end
-          res.delete(nil)
-          res
-        )
-      end
-
-      # TODO: parse options
-      # @param [String] col +COL_NAME:TYPE:{OPTIONS}
-      def parse_column(col)
-        sc = StringScanner.new(col)
-        name = sc.scan(/[^:]+/)
-        sc.scan(/:/)
-        type = sc.scan(/[^:]+/)
-        sc.scan(/:/)
-        options = sc.scan(/\{.+\}$/)
-
-        method = name[0]
-        name = name[1..-1]
-        {name: name, type: type, options: options, method: method}
-      end
 
       # @param [String] col +COL_NAME:TYPE:{OPTIONS}
       def to_method(table, col)
