+++ lib/mi.rb
@@ -1,4 +1,7 @@
 require "mi/version"
+require 'generators/mi'
+require 'generators/mi/mi_generator'
+require 'generators/mi/create_generator'
 
 module Mi
   # Your code goes here...
